server/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-extensions==3.2.3
djangorestframework==3.15.1
psycopg2==2.9.9
pydot==2.0.0
pyparsing==3.1.1
sqlparse==0.4.4
tzdata==2024.1
